refactor: move dashboard specific files to nested subfolders and update code
