+++ gui/dashboards/hashtags/__init__.py
@@ -1,0 +1,3 @@
+from .dashboard import HashtagsDashboardPage
+
+__all__ = ["HashtagsDashboardPage"]
