+++ gui/dashboards/ngrams/__init__.py
@@ -1,0 +1,3 @@
+from .dashboard import NgramsDashboardPage
+
+__all__ = ["NgramsDashboardPage"]
