+++ gui/dashboards/ngrams/dashboard.py
@@ -20,14 +20,11 @@ from analyzers.ngrams.ngrams_stats.interface import (
     OUTPUT_NGRAM_STATS,
 )
 from analyzers.ngrams.ngrams_stats.interface import interface as ngram_stats_interface
-from analyzers.ngrams.ngrams_web.plots import (
-    SamplingMetadata,
-    plot_scatter_echart,
-    sample_ngram_data,
-)
 from gui.session import GuiSession
 
-from .base_dashboard import BaseDashboardPage
+from ..base_dashboard import BaseDashboardPage
+from .data import filter_ngrams_by_text, make_detail_columns, make_summary_columns
+from .plots import SamplingMetadata, plot_scatter_echart, sample_ngram_data
 
 
 class NgramsDashboardPage(BaseDashboardPage):
@@ -49,23 +46,19 @@ class NgramsDashboardPage(BaseDashboardPage):
 
     def __init__(self, session: GuiSession):
         super().__init__(session=session)
-        # State for tracking selection
         self._selected_words: str | None = None
         self._selected_series_index: int | None = None
         self._selected_data_index: int | None = None
 
-        # State for filter
         self._filter_text: str | None = None
         self._filter_applied: bool = False
         self._all_ngram_options: list[str] = []
 
-        # DataFrames
         self._df_stats: pl.DataFrame | None = None
         self._df_full: pl.DataFrame | None = None
         self._df_stats_sampled: pl.DataFrame | None = None
         self._sampling_metadata: SamplingMetadata | None = None
 
-        # UI component references (set during render)
         self._chart: ui.echart | None = None
         self._grid: ui.aggrid | None = None
         self._info_label: ui.label | None = None
@@ -78,84 +71,27 @@ class NgramsDashboardPage(BaseDashboardPage):
         self._show_all_btn: ui.button | None = None
 
     def _get_top_n_summary(self, n: int = 100) -> pl.DataFrame:
-        """
-        Get top N n-grams by frequency for the summary view.
-
-        Respects the current filter text to show only matching n-grams.
-
-        Args:
-            n: Number of top n-grams to return (default 100)
-
-        Returns:
-            DataFrame with summary columns, sorted by total_reps descending
-        """
-        # Use filtered stats to respect current filter
         df = self._get_filtered_stats()
         if df.is_empty():
             return pl.DataFrame()
 
         return (
-            df.select(
-                [
-                    COL_NGRAM_WORDS,
-                    COL_NGRAM_TOTAL_REPS,
-                    COL_NGRAM_DISTINCT_POSTER_COUNT,
-                ]
-            )
-            .sort(COL_NGRAM_TOTAL_REPS, descending=True)
-            .head(n)
-            .rename(
-                {
-                    COL_NGRAM_WORDS: "N-gram content",
-                    COL_NGRAM_TOTAL_REPS: "Total repetitions",
-                    COL_NGRAM_DISTINCT_POSTER_COUNT: "Unique posters",
-                }
-            )
+            make_summary_columns(df).sort("Total repetitions", descending=True).head(n)
         )
 
     def _get_filtered_full_data(self, words: str) -> pl.DataFrame:
-        """
-        Get all occurrences of a specific n-gram from the full report.
-
-        Args:
-            words: The n-gram words string to filter by
-
-        Returns:
-            DataFrame with detail columns for the selected n-gram
-        """
         if self._df_full is None or self._df_full.is_empty():
             return pl.DataFrame()
 
-        return (
-            self._df_full.filter(pl.col(COL_NGRAM_WORDS) == words)
-            .select(
-                [
-                    COL_AUTHOR_ID,
-                    COL_NGRAM_WORDS,
-                    COL_MESSAGE_TEXT,
-                    COL_MESSAGE_TIMESTAMP,
-                ]
-            )
-            .with_columns(
-                pl.col(COL_MESSAGE_TIMESTAMP).dt.strftime("%B %d, %Y %I:%M %p")
-            )
-            .rename(
-                {
-                    COL_AUTHOR_ID: "User ID",
-                    COL_NGRAM_WORDS: "N-gram content",
-                    COL_MESSAGE_TEXT: "Post content",
-                    COL_MESSAGE_TIMESTAMP: "Timestamp",
-                }
-            )
+        return self._df_full.filter(pl.col(COL_NGRAM_WORDS) == words).pipe(
+            make_detail_columns
         )
 
     def _update_info_label(self) -> None:
-        """Update the info label based on current selection and filter state."""
         if self._info_label is None:
             return
 
         if self._selected_words is not None:
-            # Show detail for selected n-gram (click takes precedence)
             if self._df_full is not None:
                 count = self._df_full.filter(
                     pl.col(COL_NGRAM_WORDS) == self._selected_words
@@ -166,7 +102,6 @@ class NgramsDashboardPage(BaseDashboardPage):
                 f"N-gram: '{self._selected_words}' — {count:,} total repetitions"
             )
         elif self._filter_text:
-            # Show filter status
             df_filtered = self._get_filtered_stats()
             count = df_filtered.height
             if count == 0:
@@ -175,20 +110,17 @@ class NgramsDashboardPage(BaseDashboardPage):
                     "Try a different search term."
                 )
             elif not self._filter_applied:
-                # User is typing, show hint to press Enter
                 self._info_label.text = (
                     f"Filter: '{self._filter_text}' — {count:,} matches found. "
                     "Press Enter to apply filter to chart and grid."
                 )
             else:
-                # Filter has been applied
                 self._info_label.text = (
                     f"Showing {min(count, 100):,} of {count:,} n-grams "
                     f"matching '{self._filter_text}'. "
                     "Click a point to view all occurrences."
                 )
         else:
-            # Default summary view
             self._info_label.text = (
                 "Showing top 100 n-grams by frequency. "
                 "Type to search, then press Enter to filter. "
@@ -196,24 +128,19 @@ class NgramsDashboardPage(BaseDashboardPage):
             )
 
     def _update_grid(self) -> None:
-        """Update the data grid based on current selection state."""
         if self._grid is None:
             return
 
         if self._selected_words is None:
-            # Show summary view
             df_display = self._get_top_n_summary()
         else:
-            # Show filtered detail view
             df_display = self._get_filtered_full_data(self._selected_words)
 
         if df_display.is_empty():
             self._grid.options["rowData"] = []
             self._grid.options["columnDefs"] = []
         else:
-            # Convert to ag-grid format
             self._grid.options["rowData"] = df_display.to_dicts()
-            # Build column definitions with special handling for Post content
             column_defs = []
             for col in df_display.columns:
                 col_def = {
@@ -222,7 +149,6 @@ class NgramsDashboardPage(BaseDashboardPage):
                     "filter": True,
                     "resizable": True,
                 }
-                # Add tooltip for Post content column to show full text on hover
                 if col == "Post content":
                     col_def[":tooltipValueGetter"] = "(params) => params.value"
                 column_defs.append(col_def)
@@ -230,16 +156,6 @@ class NgramsDashboardPage(BaseDashboardPage):
         self._grid.update()
 
     def _highlight_point(self, series_index: int, data_index: int) -> None:
-        """
-        Highlight a point using ECharts dispatchAction (no chart redraw).
-
-        This is much more efficient than adding a separate highlight series,
-        as it only toggles visual styles without re-rendering the chart.
-
-        Args:
-            series_index: Index of the series containing the point
-            data_index: Index of the data point within the series
-        """
         if self._chart is None:
             return
         self._chart.run_chart_method(
@@ -248,42 +164,15 @@ class NgramsDashboardPage(BaseDashboardPage):
         )
 
     def _downplay_point(self, series_index: int, data_index: int) -> None:
-        """
-        Remove highlight from a point using ECharts dispatchAction (no chart redraw).
-
-        Args:
-            series_index: Index of the series containing the point
-            data_index: Index of the data point within the series
-        """
         self._chart.run_chart_method(
             "dispatchAction",
             {"type": "downplay", "seriesIndex": series_index, "dataIndex": data_index},
         )
 
     def _clear_all_highlights(self) -> None:
-        """
-        Clear all highlights from the chart.
-
-        Calling dispatchAction with type 'downplay' without specifying
-        seriesIndex/dataIndex will downplay all highlighted elements.
-        """
         self._chart.run_chart_method("dispatchAction", {"type": "downplay"})
 
     def _handle_point_click(self, e) -> None:
-        """
-        Handle click events on scatter plot points.
-
-        Implements toggle behavior:
-        - Click unselected point → select it
-        - Click selected point → deselect it
-        - Click different point → switch selection
-
-        Uses ECharts dispatchAction for efficient highlight/downplay without
-        triggering a full chart redraw.
-
-        Args:
-            e: Click event object with point data (EChartPointClickEventArguments)
-        """
         clicked_words = e.data.get("words")
         if clicked_words is None:
             return
@@ -291,7 +180,6 @@ class NgramsDashboardPage(BaseDashboardPage):
         series_index = e.series_index
         data_index = e.data_index
 
-        # Downplay previous selection if any
         if (
             self._selected_series_index is not None
             and self._selected_data_index is not None
@@ -299,120 +187,62 @@ class NgramsDashboardPage(BaseDashboardPage):
             self._downplay_point(self._selected_series_index, self._selected_data_index)
 
         if self._selected_words == clicked_words:
-            # Clicking same point - deselect (already downplayed above)
             self._selected_words = None
             self._selected_series_index = None
             self._selected_data_index = None
         else:
-            # Clicking new point - highlight it
             self._selected_words = clicked_words
             self._selected_series_index = series_index
             self._selected_data_index = data_index
             self._highlight_point(series_index, data_index)
 
-        # Update UI components
         self._update_info_label()
         self._update_grid()
 
     def _handle_filter_change(self, e) -> None:
-        """
-        Handle n-gram filter input changes (fires on every keystroke).
-
-        Updates info label to guide user.
-        Does NOT update chart/grid (expensive operations) - those happen on Enter.
-
-        Args:
-            e: Change event from ui.input with value attribute
-        """
         self._filter_text = e.value if e.value else None
         self._filter_applied = False
-        # Update info label to show hint
         self._update_info_label()
 
     def _handle_enter_press(self, e) -> None:
-        """
-        Handle Enter key press in search input.
-
-        Updates the expensive visualizations (chart and grid) with the
-        current filter text. This provides good performance by avoiding
-        continuous redraws on every keystroke.
-
-        Args:
-            e: Keydown event from ui.input
-        """
-        # Clear any previous selection
         self._selected_words = None
         self._selected_series_index = None
         self._selected_data_index = None
         self._clear_all_highlights()
 
-        # Mark filter as applied
         self._filter_applied = True
 
-        # Update expensive visualizations
         self._update_chart_with_filter()
         self._update_grid()
-
-        # Update info label to show results
         self._update_info_label()
 
     def _get_filtered_stats(self) -> pl.DataFrame:
-        """
-        Return a filtered view of the n-gram stats.
-
-        - No active filter  → returns self._df_stats_sampled (fast default view)
-        - Active filter     → searches self._df_stats (full data) so users can
-                              find any n-gram, not just those in the sampled view.
-        """
         if self._df_stats is None:
             return pl.DataFrame()
 
         if not self._filter_text:
-            # Default: return sampled data for performance
             return (
                 self._df_stats_sampled
                 if self._df_stats_sampled is not None
                 else self._df_stats
             )
 
-        # Filter: always search full dataset
-        return self._df_stats.filter(
-            pl.col(COL_NGRAM_WORDS).str.contains(f"(?i){self._filter_text}")
-        )
+        return filter_ngrams_by_text(self._df_stats, self._filter_text)
 
     def _handle_clear(self) -> None:
-        """
-        Handle clear button click on search input.
-
-        Resets the chart and grid to show the initial unfiltered state.
-        """
-        # Clear filter state
         self._filter_text = None
         self._filter_applied = False
 
-        # Clear any selection
         self._selected_words = None
         self._selected_series_index = None
         self._selected_data_index = None
         self._clear_all_highlights()
 
-        # Re-render chart with full dataset
         self._update_chart_with_filter()
-
-        # Update grid and info label
         self._update_grid()
         self._update_info_label()
 
     async def _load_and_render_async(self) -> None:
-        """
-        Load parquet data and build the chart, fully off the main thread.
-
-        Call order:
-            1. run.io_bound  → read parquet files
-            2. run.cpu_bound → sample data          (plots.sample_ngram_data)
-            3. run.cpu_bound → build ECharts option (plots.plot_scatter_echart)
-            4. Main thread   → push to UI
-        """
         stats_path = self.get_output_parquet_path(OUTPUT_NGRAM_STATS)
         full_path = self.get_output_parquet_path(OUTPUT_NGRAM_FULL)
 
@@ -423,7 +253,6 @@ class NgramsDashboardPage(BaseDashboardPage):
                 )
             return
 
-        # --- Step 1: I/O-bound parquet reads ---
         try:
             self._df_stats = await run.io_bound(pl.read_parquet, stats_path)
             if full_path:
@@ -440,7 +269,6 @@ class NgramsDashboardPage(BaseDashboardPage):
                 self._show_error(self._chart_loading, "No n-gram data available.")
             return
 
-        # --- Steps 2 & 3: CPU-bound sampling + chart building ---
         try:
             self._df_stats_sampled, self._sampling_metadata = await run.cpu_bound(
                 sample_ngram_data,
@@ -457,7 +285,6 @@ class NgramsDashboardPage(BaseDashboardPage):
                 self._show_error(self._chart_loading, f"Could not build chart: {exc}")
             return
 
-        # --- Step 4: Populate autocomplete from FULL data ---
         if self._ngram_select is not None:
             self._all_ngram_options = (
                 self._df_stats.select(pl.col(COL_NGRAM_WORDS).unique())
@@ -467,7 +294,6 @@ class NgramsDashboardPage(BaseDashboardPage):
             )
             self._ngram_select.set_autocomplete(self._all_ngram_options)
 
-        # --- Step 5: Push to UI (main thread) ---
         if (
             self._chart is None
             or self._chart_content is None
@@ -485,7 +311,6 @@ class NgramsDashboardPage(BaseDashboardPage):
             self._show_content(self._grid_loading, self._grid_content)
 
     def _update_sampling_info_label(self) -> None:
-        """Update the sampling label and show/hide the 'Show all' button."""
         if self._sampling_label is None or self._sampling_metadata is None:
             return
         self._sampling_label.text = self._sampling_metadata.sampling_message
@@ -493,7 +318,6 @@ class NgramsDashboardPage(BaseDashboardPage):
             self._show_all_btn.set_visibility(self._sampling_metadata.is_sampled)
 
     async def _handle_show_all_click(self) -> None:
-        """Show confirmation dialog for very large datasets, then load all data."""
         if self._df_stats is None:
             return
 
@@ -518,7 +342,6 @@ class NgramsDashboardPage(BaseDashboardPage):
             await self._load_full_dataset()
 
     async def _load_full_dataset(self) -> None:
-        """Rebuild the chart from the complete (unsampled) dataset."""
         if self._show_all_btn is not None:
             self._show_all_btn.set_enabled(False)
         if self._sampling_label is not None:
@@ -555,7 +378,6 @@ class NgramsDashboardPage(BaseDashboardPage):
         self._update_sampling_info_label()
 
     def _update_chart_with_filter(self) -> None:
-        """Re-render the chart with the currently filtered data."""
         if self._chart is None:
             return
 
@@ -566,13 +388,11 @@ class NgramsDashboardPage(BaseDashboardPage):
             self._chart.update()
             return
 
-        # Filter results are already a small subset — run synchronously
         option = plot_scatter_echart(df_filtered, enable_large_mode=False)
         self._chart.options.update(option)
         self._chart.update()
 
     def render_content(self) -> None:
-        """Render the scatter plot and data grid with loading states."""
         ui.add_css(
             """
             .ag-tooltip {
@@ -584,7 +404,6 @@ class NgramsDashboardPage(BaseDashboardPage):
         )
         with ui.row().classes("w-full justify-center"):
             with ui.column().classes("w-3/4 q-pa-md gap-4"):
-                # Scatter plot card
                 with ui.card().classes("w-full"):
                     self._ngram_select = (
                         ui.input(
@@ -598,7 +417,6 @@ class NgramsDashboardPage(BaseDashboardPage):
                         .on("clear", self._handle_clear)
                     )
 
-                    # Loading / chart containers
                     self._chart_loading, self._chart_content = (
                         self._create_loading_container("500px")
                     )
@@ -609,7 +427,6 @@ class NgramsDashboardPage(BaseDashboardPage):
                             .style("height: 500px")
                         )
 
-                # Sampling info row (hidden until data loads)
                 with ui.row().classes("w-full items-center gap-4"):
                     self._sampling_label = ui.label("").classes(
                         "text-body2 text-grey-7"
@@ -621,14 +438,12 @@ class NgramsDashboardPage(BaseDashboardPage):
                     ).props("outline dense")
                     self._show_all_btn.set_visibility(False)
 
-                # Data viewer card
                 with ui.card().classes("w-full"):
                     with ui.card_section():
                         ui.label("Data viewer").classes("text-h6")
                     self._info_label = ui.label("Loading data...").classes(
                         "text-body2 text-grey-7 q-mb-sm"
                     )
-                    # Loading / grid containers
                     self._grid_loading, self._grid_content = (
                         self._create_loading_container("400px")
                     )
@@ -652,5 +467,4 @@ class NgramsDashboardPage(BaseDashboardPage):
                             .style("height: 400px")
                         )
 
-        # Trigger async load after page renders
         ui.timer(0, self._load_and_render_async, once=True)
