+++ gui/dashboards/placeholder/__init__.py
@@ -1,0 +1,3 @@
+from .dashboard import PlaceholderDashboard
+
+__all__ = ["PlaceholderDashboard"]
